+++ node/http_utils.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"github.com/gorilla/mux"
 	"github.com/opentracing/opentracing-go"
-	"github.com/rovergulf/rbn/core"
 	"github.com/rovergulf/rbn/pkg/resutil"
 	"go.uber.org/zap"
 	"net/http"
@@ -91,22 +90,4 @@ func (n *Node) httpResponse(w http.ResponseWriter, i interface{}, statusCode ...
 		w.WriteHeader(http.StatusInternalServerError)
 		w.Write([]byte(fmt.Sprintf("Unable to write json response: %s", err)))
 	}
-}
-
-type StatusRes struct {
-	LastHash   string                   `json:"block_hash,omitempty" yaml:"last_hash,omitempty"`
-	Number     uint64                   `json:"chain_length,omitempty" yaml:"chain_length,omitempty"`
-	KnownPeers map[string]PeerNode      `json:"peers_known,omitempty" yaml:"known_peers,omitempty"`
-	PendingTXs map[string]core.SignedTx `json:"pending_txs,omitempty" yaml:"pending_t_xs,omitempty"`
-	IsMining   bool                     `json:"is_mining" yaml:"is_mining"`
-	DbSize     map[string]int64         `json:"db_size" yaml:"db_size"`
-}
-
-type SyncRes struct {
-	Blocks []*core.Block `json:"blocks" yaml:"blocks"`
-}
-
-type AddPeerRes struct {
-	Success bool   `json:"success" yaml:"success"`
-	Error   string `json:"error,omitempty" yaml:"error,omitempty"`
 }